odoo-addons/13.0/requirements.txt
graphene
graphql-server-core
pyjwt